Split subscribers
